+++ src/execute/load_branches.go
@@ -21,7 +21,7 @@ func LoadBranches(args LoadBranchesArgs) (domain.Branches, domain.BranchesSnapsh
 		}
 		exit, err := validate.HandleUnfinishedState(validate.UnfinishedStateArgs{
 			Connector:               nil,
-			Debug:                   args.Debug,
+			Verboe:                  args.Verbose,
 			InitialBranchesSnapshot: branchesSnapshot,
 			InitialConfigSnapshot:   args.Repo.ConfigSnapshot,
 			InitialStashSnapshot:    stashSnapshot,
@@ -82,7 +82,7 @@ func LoadBranches(args LoadBranchesArgs) (domain.Branches, domain.BranchesSnapsh
 type LoadBranchesArgs struct {
 	Repo                  *OpenRepoResult
 	Fetch                 bool
-	Debug                 bool
+	Verbose               bool
 	HandleUnfinishedState bool
 	Lineage               config.Lineage
 	PushHook              bool