+++ src/subshell/backend_runner.go
@@ -86,7 +86,7 @@ OUTPUT END
 
 func printHeader(cmd string, args ...string) {
 	quoted := stringslice.SurroundEmptyWith(args, `"`)
-	text := "\n(debug) " + cmd + " " + strings.Join(quoted, " ")
+	text := "\n(verbose) " + cmd + " " + strings.Join(quoted, " ")
 	_, err := color.New(color.Bold).Println(text)
 	if err != nil {
 		fmt.Println(text)